Load locations only for visible nodes in paged list
