+++ src/components/Stream/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo, useRef } from 'react'
+import React, { useEffect, useMemo } from 'react'
 import { useParams } from 'react-router-dom'
 
 import { useStore } from '../../contexts/Store'
@@ -78,27 +78,6 @@ const ActiveNodeSetter = ({ id }: NodeProps) => {
   return null
 }
 
-const ReverseGeoCodingLoader = () => {
-  const { visibleNodes } = useStore()
-  const { loadNodeLocations } = useController()
-  const fetching = useRef(false)
-
-  useEffect(() => {
-    const nodesWithoutLocation = visibleNodes
-      .filter(({ location }) => location && !location.isReverseGeoCoded)
-
-    if (nodesWithoutLocation.length > 0 && !fetching.current) {
-      fetching.current = true
-      loadNodeLocations(nodesWithoutLocation)
-        .then(() => {
-          fetching.current = false
-        })
-    }
-  }, [visibleNodes, loadNodeLocations])
-
-  return null
-}
-
 interface ParamTypes {
   nodeId: string
   streamId: string
@@ -121,7 +100,6 @@ export default () => {
       <StreamLoader id={streamId} />
       <ActiveNodeSetter id={nodeId} />
       <TopologyList id={streamId} />
-      <ReverseGeoCodingLoader />
     </>
   )
 }
